api/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
drf-yasg
drf-yasg[validation]